BUILD: Cache the last working Boost library suffix
The boost libraries can have an optional suffix, which can be any
combination of "mt" (for multi-threaded), the compiler tag, flags for
static, debug, etc. Because we need to be able to find all of them,
the library checks in boost.m4 tries them all in sequence.

Unfortunately, boost.m4 retries each combination for every single
library. This is slow.

However, the suffix is usually the same among all libraries of a given
boost installation. So now, once we have found a working library, we
simply remember the suffix of this library. And for every subsequent
check, we try this suffix first.

This essentially speeds up boost.m4 considerably, because from the
second library check onwards, only a single library needs to be
checked.